refactor: swagger api
