+++ packages/nocodb/src/lib/services/index.ts
@@ -30,3 +30,4 @@ export * as dataService from './dataService';
 export * as bulkDataService from './dataService/bulkData';
 export * as cacheService from './cacheService';
 export * as auditService from './auditService';
+export * as swaggerService from './swaggerService';
