+++ packages/nocodb/src/lib/services/swaggerService/templates/params.ts
@@ -1,6 +1,6 @@
 import { SwaggerColumn } from '../getSwaggerColumnMetas';
 import { RelationTypes, UITypes } from 'nocodb-sdk';
-import LinkToAnotherRecordColumn from '../../../../../models/LinkToAnotherRecordColumn';
+import LinkToAnotherRecordColumn from '../../../models/LinkToAnotherRecordColumn';
 
 export const rowIdParam = {
   schema: {
