+++ packages/nocodb/src/lib/services/swaggerService/templates/paths.ts
